Refactor api's underlying runtime to reflect the layout of the preamp's registers
This adds a little complexity to the set_group and set_zone api functions but makes it so that the runtime doesn't need to know everything. This will need some testing and probably a second update once we know what sequence the register needs to be set in to avoid obnoxious audio pops, but for now it should be good enough to start playing with.