Add test parameter
This will allow us to test a url in the future wuthout playing it out an output. While adding this parameter the argument handling was upgraded to use the standard python argument parser argparse. In the process the alsa device parameter was removed since it can easily be derived from the source id.