Make SSH setup a little better
Mark had an issue where a special ssh key (the last one generated) was being copied and ssh wouldn't use it. In an attempt to avoid this we check for the default keys and generate a new default key if none exist. The copy is also fixed to only execute when the passwordless ssh attempt fails as was originally intended.