Use exclude_none instead of exclude_unset
unfiltered notes:
This fixes some weird issues that happened with fields for handling comparisons in test environment. exclude_none matches our intention better anyway, we don't want to report fields that are unpopulated (they are used for differentiation between similar models). Our models do not include None and pydantic uses None to indicate a field that is not yet set.